Finalizing cluster rename
